Change test type and add template for Function Question
This is a change I've been trying to avoid for a while, but it's clear that in
order to allow editing of questions, we need to keep the test function as
a string. This also has the benefit of bypassing the non-generic
restriction imposed by JLD. On the other hand, we now have no
guarantee that the test is a valid function --- this should fail quickly
though.